refactor: update file structure
