+++ src/handlers/handleCommandInput.js
@@ -1,0 +1,11 @@
+import { parseCommandLine } from '../utils/parseCommandLine.js';
+import { executeCommand } from './executeCommand.js';
+
+export const handleCommandInput = async (input) => {
+  try {
+    let commandObj = parseCommandLine(input);
+    await executeCommand(commandObj);
+  } catch (err) {
+    console.log('Operation failed');
+  }
+};
